Create Printing out
Printing out
Printing text is one of the first steps in Java programming. It helps display messages, debug code, and interact with users.

Complete the program to print a welcome message!

Instructions

Replace ____ with the correct Java keyword for printing - println.
Tip: you can view the Slides in the output window. On mobile, click on the three horizontal lines.