Create Printing as a calculator
Printing as a calculator
Java can be used for quick calculations, just like a calculator! Instead of manually computing values, you can use System.out.println() to display the result of an arithmetic operation.

Complete the program to print the sum of two numbers.

Instructions

Print the result of 1+3 to the output.