Create Creating a variable
Creating a variable
Before using a variable, you need to declare it by specifying its type and name. Java requires every variable to have a type that defines the kind of data it holds.

Instructions

Create a String variable called language that has the value of "Java".
Create a variable of type int called yearOfOrigin, that is set to 1995.