Create Strings, chars, and booleans
Strings, chars, and booleans
While the String data type is not a primitive, it is super important as well. It's time to practice them along with char and boolean.

Instructions

Declare the tshirtSize variable using the correct primitive type.
Assign true to the isOnSale variable.
Declare String variable called color, setting its value to "green".