Create How long is a String?
How long is a String?
String is a collection of characters and oftentimes, you'll need to find its length. Instead of counting though, there is the .length() method!

Instructions

Use the .length() method to find out the length of the bookName variable, saving the result as bookNameLength.
Find the length of the author's name, "Terry Pratchett", and save it using the correct primitive type.