Create Nobel winners in a row
Nobel winners in a row
Arrays are very useful for storing values of the same type. They can help you keep information in order, such as the chronological order of your favorite Nobel Prize laureates for Physics!

Instructions

Create a String array nobelLaureates.
Assign the values of 1903, 1921, and 1938 into the array years.