Create Nobel winners revisited
Nobel winners revisited
Remember the Nobel Prize laureates for Physics from the last exercise? You decide to prefer Heisenberg to Fermi, so you want to update your records.

Instructions

Change the third element of nobelLaureates to "Heisenberg".
Change the third element of years to 1932.
Finish the print statement to print the last element of both arrays.