Create Equal or not
Equal or not
Checking whether values are equal or not is a useful skill to have. Do you think the following are equal or not? Sometimes things are not as they seem in Java!

Instructions

Save the result of checking whether 193 is equal to the value of a to the variable first.
Complete the print statement to display the variable first.